// src/types/appraisalView.ts
// src/types/appraisalView.ts

export enum AppraisalView {
    LIST_QUESTIONS = 'LIST_QUESTIONS',
    CREATE_QUESTION = 'CREATE_QUESTION',
    VIEW_QUESTION = 'VIEW_QUESTION',
    EDIT_QUESTION = 'EDIT_QUESTION',
    LIST_METRICS = 'LIST_METRICS',
    CREATE_METRIC = 'CREATE_METRIC',
    VIEW_METRIC = 'VIEW_METRIC',
    EDIT_METRIC = 'EDIT_METRIC',
  }
  